requirements.txt
matplotlib===3.10.8
pygame-ce===2.5.7
numpy===2.4.2